+++ packages/provider/test/multi_provider_test.dart
@@ -65,47 +65,65 @@ void main() {
 
       // p1 cannot access to p1/p2/p3
       expect(
-        () => Provider.of<int>(k1.currentContext),
+        () => Provider.of<int>(k1.currentContext, listen: false),
         throwsProviderNotFound(
-            valueType: int, widgetType: _typeOf<Provider<int>>()),
+          valueType: int,
+          widgetType: _typeOf<Provider<int>>(),
+        ),
       );
       expect(
-        () => Provider.of<String>(k1.currentContext),
+        () => Provider.of<String>(k1.currentContext, listen: false),
         throwsProviderNotFound(
-            valueType: String, widgetType: _typeOf<Provider<int>>()),
+          valueType: String,
+          widgetType: _typeOf<Provider<int>>(),
+        ),
       );
       expect(
-        () => Provider.of<double>(k1.currentContext),
+        () => Provider.of<double>(k1.currentContext, listen: false),
         throwsProviderNotFound(
-            valueType: double, widgetType: _typeOf<Provider<int>>()),
+          valueType: double,
+          widgetType: _typeOf<Provider<int>>(),
+        ),
       );
 
       // p2 can access only p1
-      expect(Provider.of<int>(k2.currentContext), 42);
+      expect(Provider.of<int>(k2.currentContext, listen: false), 42);
       expect(
-        () => Provider.of<String>(k2.currentContext),
+        () => Provider.of<String>(k2.currentContext, listen: false),
         throwsProviderNotFound(
-            valueType: String, widgetType: _typeOf<Provider<String>>()),
+          valueType: String,
+          widgetType: _typeOf<Provider<String>>(),
+        ),
       );
       expect(
-        () => Provider.of<double>(k2.currentContext),
+        () => Provider.of<double>(k2.currentContext, listen: false),
         throwsProviderNotFound(
-            valueType: double, widgetType: _typeOf<Provider<String>>()),
+          valueType: double,
+          widgetType: _typeOf<Provider<String>>(),
+        ),
       );
 
       // p3 can access both p1 and p2
-      expect(Provider.of<int>(k3.currentContext), 42);
-      expect(Provider.of<String>(k3.currentContext), 'foo');
+      expect(Provider.of<int>(k3.currentContext, listen: false), 42);
       expect(
-        () => Provider.of<double>(k3.currentContext),
+        Provider.of<String>(k3.currentContext, listen: false),
+        'foo',
+      );
+      expect(
+        () => Provider.of<double>(k3.currentContext, listen: false),
         throwsProviderNotFound(
-            valueType: double, widgetType: _typeOf<Provider<double>>()),
+          valueType: double,
+          widgetType: _typeOf<Provider<double>>(),
+        ),
       );
 
       // the child can access them all
-      expect(Provider.of<int>(keyChild.currentContext), 42);
-      expect(Provider.of<String>(keyChild.currentContext), 'foo');
-      expect(Provider.of<double>(keyChild.currentContext), 44);
+      expect(Provider.of<int>(keyChild.currentContext, listen: false), 42);
+      expect(
+        Provider.of<String>(keyChild.currentContext, listen: false),
+        'foo',
+      );
+      expect(Provider.of<double>(keyChild.currentContext, listen: false), 44);
     });
   });
 }